- change package structure - add go.mod in all package
