+++ cmd/validate.go
@@ -1,4 +1,4 @@
-package core
+package cmd
 
 import (
 	"github.com/spf13/cobra"
