--- providers/aws/eks/ecr.go
@@ -1,1 +1,0 @@
-package eks
