+++ providers/aws/go.mod
@@ -1,0 +1,12 @@
+module github.com/Qovery/pleco/providers/aws
+
+go 1.16
+
+require (
+	github.com/aws/aws-sdk-go v1.38.1
+	github.com/prometheus/common v0.19.0
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/cobra v1.1.3
+	k8s.io/client-go v0.20.5
+	sigs.k8s.io/aws-iam-authenticator v0.5.2
+)
