+++ providers/aws/utils.go
@@ -1,4 +1,4 @@
-package utils
+package aws
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
@@ -7,7 +7,6 @@ import (
 	"time"
 )
 
-
 func GetTimeInfos(tags []*ec2.Tag) (time.Time, int64) {
 	var creationDate = time.Now()
 	var ttl int64
@@ -15,7 +14,7 @@ func GetTimeInfos(tags []*ec2.Tag) (time.Time, int64) {
 	for i := range tags {
 		if *tags[i].Key == "creationDate" {
 			creationTime, _ := strconv.ParseInt(*tags[i].Value, 10, 64)
-			creationDate = time.Unix(creationTime,0)
+			creationDate = time.Unix(creationTime, 0)
 		}
 		if *tags[i].Key == "ttl" {
 			result, _ := strconv.ParseInt(*tags[i].Value, 10, 64)
@@ -31,18 +30,18 @@ func CheckIfExpired(creationTime time.Time, ttl int64) bool {
 	return time.Now().After(expirationTime)
 }
 
-func AddCreationDateTag (ec2Session ec2.EC2, idsToTag []*string, creationDate time.Time, ttl int64) error {
+func AddCreationDateTag(ec2Session ec2.EC2, idsToTag []*string, creationDate time.Time, ttl int64) error {
 	_, err := ec2Session.CreateTags(
 		&ec2.CreateTagsInput{
-			Resources: 	idsToTag,
+			Resources: idsToTag,
 			Tags: []*ec2.Tag{
 				{
-					Key: aws.String("creationDate"),
+					Key:   aws.String("creationDate"),
 					Value: aws.String(creationDate.String()),
 				},
 				{
-					Key: aws.String("ttl"),
-					Value: aws.String(strconv.FormatInt(ttl,10)),
+					Key:   aws.String("ttl"),
+					Value: aws.String(strconv.FormatInt(ttl, 10)),
 				},
 			},
 		})
