+++ providers/k8s/go.mod
@@ -1,0 +1,11 @@
+module github.com/Qovery/pleco/providers/k8s
+
+go 1.16
+
+require (
+	github.com/aws/aws-sdk-go v1.38.1
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/cobra v1.1.3
+	k8s.io/apimachinery v0.20.5
+	k8s.io/client-go v0.20.5
+)
