+++ providers/k8s/namespaces.go
@@ -3,7 +3,6 @@ package k8s
 import (
 	"context"
 	"fmt"
-	"github.com/Qovery/pleco/utils"
 	log "github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -12,17 +11,17 @@ import (
 )
 
 type kubernetesNamespace struct {
-	Name string
+	Name                string
 	NamespaceCreateTime time.Time
-	Status string
-	TTL int64
+	Status              string
+	TTL                 int64
 }
 
 func listTaggedNamespaces(clientSet *kubernetes.Clientset, tagName string) ([]kubernetesNamespace, error) {
 	var taggedNamespaces []kubernetesNamespace
 
 	listOptions := metav1.ListOptions{
-		LabelSelector:       tagName,
+		LabelSelector: tagName,
 	}
 
 	log.Debugf("Listing all Kubernetes namespaces with %s label", tagName)
@@ -89,7 +88,7 @@ func DeleteExpiredNamespaces(clientSet *kubernetes.Clientset, tagName string, dr
 	}
 
 	for _, namespace := range namespaces {
-		if utils.CheckIfExpired(namespace.NamespaceCreateTime, namespace.TTL) {
+		if CheckIfExpired(namespace.NamespaceCreateTime, namespace.TTL) {
 			err := deleteNamespace(clientSet, namespace, dryRun)
 			if err != nil {
 				log.Errorf("error while trying to delete namespace: %s", err)
